requirements-dev.txt
# The following are only required for Cement development
setuptools>=39.2.0
wheel>=0.31.0
pytest>=4.3.1
pytest-cov>=2.6.1
coverage>=4.5.3
flake8
black
isort
mock
sphinx
sphinx_rtd_theme
recommonmark
pre-commit

# Required for optional extensions
colorlog
enlighten
ibis-framework[kerberos,impala]
impyla==0.15.0
jinja2
pylibmc
pysocks
pystache
PyYAML
redis
sasl==0.2.1
tabulate
thrift_sasl==0.2.1
thriftpy2==0.4.0
thriftpy==0.3.9
watchdog
